Add num items sold to WCRevenueStatsModel as Integer

+++ plugins/woocommerce/src/main/kotlin/org/wordpress/android/fluxc/model/WCRevenueStatsModel.kt
@@ -59,7 +59,7 @@ data class WCRevenueStatsModel(@PrimaryKey @Column private var id: Int = 0) : Id
         @SerializedName("avg_order_value")
         val avgOrderValue: Double? = null
         @SerializedName("num_items_sold")
-        val itemsSold: Double? = null
+        val itemsSold: Int? = null
     }
 
     /**
